Using the nav host to navigate between fragments
The book I am following instructs you to attach the
navController.navigate call to the button within onCreateView in the
fragment. However, this does not appear to work. After much rooting
around documents, it appears to be that the view has not attached to the
navHostFragment at this point, and so the call to view.findNavController
fails.

There are two options that I have found to fix this. The first is to use
requireActivity.findNavController(R.id.nav_host_fragment) to explictly
call the activity_main.xml layout for its nav controller. The second,
which I have gone for, is to call view.findNavController() within
onViewCreated, which is called once the view has finished creating and
attached to the root fragmentContainer.