Finishes secret message app
Uses safe args to pass information between fragments, and manipulates
the back stack to change behaviouur on using the back button.

I'm still not clear on the details of how to use Gradle, but I'll get to
that at some point; part of the issue is that I was trying to include
safeargs as a dependency as well as a plugin.

The difference between the two seems to be that pluigns can alter your
code, whereas dependencies aren't allowed to do so. Safe args creates
'Direction' classes, i.e. alters your code, and so it must be added as a
plugin. I was doing that, but *also* adding it to the dependencies,
which isn't needed.