Adds nav_graph based functionality to the custom toolbar.
Additionally, starting to figure out how the gradle build stuff works a
bit better. There appears to be a libs.toml.versions file under the
gradle folder at the project level which is the new way of managing
versions and whatnot for dependencies; this is also where the details
are going when I switch from implementation("some.sort.of.string") to
implementation(libs.string.reference).