Converts Stopwatch app to viewBindings
There's an amount of overhead to using findViewById, as the app needs to
search through the associated view hierarchy, which takes time,
especially in a deep hierarchy.

Instead, one can use veiw binding; this creates a class at compilation
time that represents the inflated view and contains properties that are
references to all objects with the view that have an id (i.e. those that
are findable through findViewById). This is then much more efficient to
reference objects.