Uses data binding for variable layout text
dataBinding and viewBinding are similar in that they both generate a
binding class for layouts, but different in when they do and how they're
used.

dataBindingg generates it only for layouts enclosed in the <layout> tag,
while viweBinding generates it for all. dataBinding allows layout
elements to access properties and methods directly from the viewModel,
whereas viewBinding requires fragment or activity code; in return,
however, it can perform more complicated logic with it.

Both have their uses and the best choice depends on the exact situation.