Alter RecyclerView adapter to use DiffUtil
When using the previous method for updating the data backing the
RecyclerView adapter, the entire list is redrawn and rebound, which is
inefficient and leaves a poor use experience when the view jumps around.

Using ListAdapter instead, which comes with its own backing list and the
ability to use DiffUtil, allows for much more efficient updating, as
only new or changed items are redrawn and rebound. It also facilitates
nicer animations.

DiffUtil has a callback to compare two items; to implement it, you
override functions to compare whether two items are the same item, and
then whether their contents are the same.